♻️ refactor(Cline): Implement insertContentTool
